chore: create a feedback form template Signed-off-by: Sonia Sandler <ssandler@redhat.com>
