+++ packages/renderer/src/lib/feedback/SendFeedback.svelte
@@ -1,22 +1,52 @@
 <script lang="ts">
+import { CloseButton, Dropdown, Modal } from '@podman-desktop/ui-svelte';
+
 import type { FeedbackCategory } from '/@api/feedback';
 
-import DevelopersFeedback from './DevelopersFeedback.svelte';
-import GitHubIssueFeedback from './GitHubIssueFeedback.svelte';
+import DevelopersFeedback from './feedbackForms/DevelopersFeedback.svelte';
+import GitHubIssueFeedback from './feedbackForms/GitHubIssueFeedback.svelte';
 
 let displayModal = false;
 const DEFAULT_CATEGORY: FeedbackCategory = 'developers';
+let category: FeedbackCategory = DEFAULT_CATEGORY;
 
-// feedback of the user
-export let category: FeedbackCategory = DEFAULT_CATEGORY;
+const FEEDBACK_CATEGORIES = new Map<FeedbackCategory, string>([
+  ['developers', '💬 Direct your words to the developers'],
+  ['feature', '🚀 Feature request'],
+  ['bug', '🪲 Bug'],
+]);
 
 window.events?.receive('display-feedback', () => {
   displayModal = true;
 });
+
+function hideModal(): void {
+  displayModal = false;
+
+  // reset fields
+  category = DEFAULT_CATEGORY;
+}
 </script>
 
-{#if category === 'developers'}
-  <DevelopersFeedback bind:displayModal={displayModal} bind:category={category} />
-{:else}
-  <GitHubIssueFeedback bind:displayModal={displayModal} bind:category={category} />
+{#if displayModal}
+  <Modal name="Share your feedback" on:close={hideModal}>
+    <div class="flex items-center justify-between pl-4 pr-3 py-3 space-x-2 text-[var(--pd-modal-header-text)]">
+      <h1 class="grow text-lg font-bold capitalize">Share your feedback</h1>
+
+      <CloseButton on:click={hideModal} />
+    </div>
+
+    <div class="relative text-[var(--pd-modal-text)] px-10 pt-4">
+      <label for="category" class="block mb-2 text-sm font-medium text-[var(--pd-modal-text)]">Category</label>
+      <Dropdown name="category" bind:value={category}
+      options={Array.from(FEEDBACK_CATEGORIES).map(e => ({ value: e[0], label: e[1] }))}>
+      </Dropdown>
+    </div>
+
+    {#if category === 'developers'}
+      <DevelopersFeedback hideModal={hideModal}/>
+    {:else}
+      <GitHubIssueFeedback hideModal={hideModal}/>
+    {/if}
+  </Modal>
 {/if}
